+++ server/src/main/java/com/codueon/boostUp/domain/lesson/repository/LessonRepositoryImpl.java
@@ -1,10 +1,10 @@
 package com.codueon.boostUp.domain.lesson.repository;
 
-import com.codueon.boostUp.domain.lesson.dto.Get.GetLesson;
-import com.codueon.boostUp.domain.lesson.dto.Get.GetMainPageLesson;
-import com.codueon.boostUp.domain.lesson.dto.Get.QGetLesson;
-import com.codueon.boostUp.domain.lesson.dto.Get.QGetMainPageLesson;
-import com.codueon.boostUp.domain.lesson.dto.Post.PostSearchLesson;
+import com.codueon.boostUp.domain.lesson.dto.get.GetLesson;
+import com.codueon.boostUp.domain.lesson.dto.get.GetMainPageLesson;
+import com.codueon.boostUp.domain.lesson.dto.get.QGetLesson;
+import com.codueon.boostUp.domain.lesson.dto.get.QGetMainPageLesson;
+import com.codueon.boostUp.domain.lesson.dto.post.PostSearchLesson;
 import com.codueon.boostUp.domain.lesson.entity.AddressInfo;
 import com.codueon.boostUp.domain.lesson.entity.LanguageInfo;
 import com.querydsl.core.BooleanBuilder;
